+++ ginga/qtw/ImageViewQt.py
@@ -871,7 +871,7 @@ class ScrolledView(QtGui.QAbstractScrollArea):
 
         self.viewer.add_callback('redraw', self._calc_scrollbars)
         self.viewer.add_callback('limits-set',
-                                 lambda v, l: self._calc_scrollbars(v))
+                                 lambda v, l: self._calc_scrollbars(v, 0))
 
     def get_widget(self):
         return self
@@ -887,11 +887,11 @@ class ScrolledView(QtGui.QAbstractScrollArea):
 
         self.v_w.resize(width, height)
 
-    def _calc_scrollbars(self, viewer):
+    def _calc_scrollbars(self, viewer, whence):
         """Calculate and set the scrollbar handles from the pan and
         zoom positions.
         """
-        if self._scrolling:
+        if self._scrolling or whence > 0:
             return
 
         # flag that suppresses a cyclical callback